Removing crappy sql folder. Updating shell script.
The one sql folder to rule them all is not suited for multiple projects.
So I removed it and changed the script to use a path relative to the
folder where the script is called.
By default it is db/migration and I call it from the root of my project.
Also hardcoded a location for the flyway installation.
Feel free to change both to suit your needs. ;)